pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet Project num.2"
authors = ["Sesiunina Olga <sesyunina.olga7@gmail.com>"]
readme = "README.md"

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.generate_differance:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
